chore: codebase refactoring

+++ common/src/main/java/com/stevekung/fishofthieves/blockentity/FishPlaqueBlockEntity.java
@@ -134,7 +134,7 @@ public class FishPlaqueBlockEntity extends BlockEntity
 
     public float getAnimation(float partialTick)
     {
-        return this.isAnimating ? (float) this.animationTickCount + partialTick : (float) this.animationTickCount;
+        return this.isAnimating ? this.animationTickCount + partialTick : (float) this.animationTickCount;
     }
 
     @Nullable
